Create a scene map for LBA1
