+++ src/resources/parsers/bricks.ts
@@ -1,29 +1,7 @@
 import { Resource } from '../load';
 import { bits } from '../../utils';
 
-const parseSceneMap = (resource: Resource, index: number) => {
-    const buffer = resource.getEntry(index);
-    const data = new DataView(buffer);
-    let offset = 0;
-    const map = [];
-
-    while (true) {
-        const opcode = data.getUint8(offset);
-        const sceneIndex = data.getUint8(offset + 1);
-        offset += 2;
-        if (opcode === 0) {
-            break;
-        }
-
-        map.push({
-            isIsland: opcode === 2,
-            index: sceneIndex,
-        });
-    }
-    return map;
-};
-
-const parseBrick = (resource: Resource, index: number) => {
+export const parseBrick = (resource: Resource, index: number) => {
     const dataView = new DataView(resource.getEntry(index));
     const height = dataView.getUint8(1);
     const offsetX = dataView.getUint8(2);
@@ -61,5 +39,3 @@ const parseBrick = (resource: Resource, index: number) => {
     }
     return pixels;
 };
-
-export { parseSceneMap, parseBrick };
