+++ www/resources/lba1.json
@@ -16,6 +16,7 @@
         { "strategy": "STATIC", "type": "BRK", "id": "BRICKS",      "description": "Isometric Bricks",      "path": "LBA/LBA_BRK.HQR" },
         { "strategy": "STATIC", "type": "BL1", "id": "LIBRARIES",   "description": "Isometric Libraries",   "path": "LBA/LBA_BLL.HQR" },
         { "strategy": "STATIC", "type": "GR1", "id": "GRIDS",       "description": "Isometric Grids",       "path": "LBA/LBA_GRI.HQR" },
+        { "strategy": "STATIC", "type": "SM1", "id": "SCENE_MAP",   "description": "Isometric Scene Map",   "index": 0 },
 
         { "strategy": "STATIC", "type": "M4A", "id": "THEME_MENU",       "description": "Adeline Theme",    "path": "LBA/MUSIC/mus_32.m4a" },
